Move the api methods from changesets_controller into the api namespaced controller
