+++ app/views/browse/changeset.html.erb
@@ -137,7 +137,7 @@
 <% end %>
 
 <div class='secondary-actions'>
-  <%= link_to(t('.changesetxml'), :controller => "changesets", :action => "show") %>
+  <%= link_to(t('.changesetxml'), :controller => "api/changesets", :action => "show") %>
   &middot;
-  <%= link_to(t('.osmchangexml'), :controller => "changesets", :action => "download") %>
+  <%= link_to(t('.osmchangexml'), :controller => "api/changesets", :action => "download") %>
 </div>
